enable top_sets parameters in dashboard
The top_sets function now has parameters for determining the lookback window for workouts to use, as well as the threshold used to determine whether a set should be considered a max-intensity lift (as opposed to a volume-day lift). This commit enables the program templates page of the shiny dashboard to use those parameters, which are at the bottom of the input sidebar.